backend/requirements.txt
Django>=5.1.5
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
PyNaCl>=1.5.0
base58>=2.1.1
python-dotenv>=1.0.1 